Extracted component LoginForm to selfcontained component. Moved redirect logic to auth saga. Exported components.
