+++ packages/react-admin/src/mui/auth/index.js
@@ -1,0 +1,2 @@
+export LoginForm from './LoginForm';
+export Logout from './Logout';
