+++ packages/react-admin/src/mui/index.js
@@ -1,3 +1,4 @@
+export * from './auth';
 export * from './button';
 export * from './detail';
 export * from './form';
