+++ packages/react-admin/src/sideEffect/saga/auth.js
@@ -1,4 +1,4 @@
-import { all, put, call, takeEvery } from 'redux-saga/effects';
+import { all, put, call, select, takeEvery } from 'redux-saga/effects';
 import { push, replace } from 'react-router-redux';
 
 import {
@@ -15,7 +15,10 @@ import {
 } from '../../actions/authActions';
 import { FETCH_ERROR } from '../../actions/fetchActions';
 import { AUTH_LOGIN, AUTH_CHECK, AUTH_ERROR, AUTH_LOGOUT } from '../../auth';
-
+const nextPathnameSelector = state => {
+    const locationState = state.routing.location.state;
+    return (locationState && locationState.pathname) || '/';
+};
 export default authClient => {
     if (!authClient) return () => null;
     function* handleAuth(action) {
@@ -33,7 +36,8 @@ export default authClient => {
                         type: USER_LOGIN_SUCCESS,
                         payload: authPayload,
                     });
-                    yield put(push(meta.pathName || '/'));
+                    const redirectTo = yield select(nextPathnameSelector);
+                    yield put(push(redirectTo));
                 } catch (e) {
                     yield put({
                         type: USER_LOGIN_FAILURE,
